finished

+++ frontend/app/lib/features/auth.js
@@ -2,17 +2,21 @@ import { createSlice } from "@reduxjs/toolkit";
 
 const getInitialState = () => {
   if (typeof window !== "undefined") {
+    // Check if localStorage is available
+    const token = localStorage.getItem("token");
+    const userId = localStorage.getItem("userId");
     return {
-      token: localStorage.getItem("token") || null,
-      userId: localStorage.getItem("userId") || null,
-      isLoggedIn: !!localStorage.getItem("token")
+      token: token || null,
+      userId: userId || null,
+      isLoggedIn: !!token,
     };
   }
 
+  // If running on the server, return default state
   return {
     token: null,
     userId: null,
-    isLoggedIn: false
+    isLoggedIn: false,
   };
 };
 
@@ -23,20 +27,26 @@ export const authSlice = createSlice({
     setLogin: (state, action) => {
       state.token = action.payload;
       state.isLoggedIn = true;
-      localStorage.setItem("token", action.payload.token);
+      if (typeof window !== "undefined") {
+        localStorage.setItem("token", action.payload);
+      }
     },
     setUserId: (state, action) => {
       state.userId = action.payload;
-      localStorage.setItem("userId", action.payload.userId);
+      if (typeof window !== "undefined") {
+        localStorage.setItem("userId", action.payload);
+      }
     },
     setLogout: (state, action) => {
       state.token = null;
       state.userId = null;
       state.isLoggedIn = false;
-      localStorage.clear();
-    }
-  }
+      if (typeof window !== "undefined") {
+        localStorage.clear();
+      }
+    },
+  },
 });
 
-export const { setLogin, setUserId, setLogout, token } = authSlice.actions;
+export const { setLogin, setUserId, setLogout } = authSlice.actions;
 export default authSlice.reducer;
